# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "A semaphore API for Distributed locks"
  version: "1.0.0"
  title: "Semaphore API"
host: "localhost:8080"
basePath: "/v1"
tags:
- name: "Semaphore"
- name: "Mutex"
schemes:
- "http"
paths:
  /semaphore/{semaphoreKey}:
    get:
      tags:
      - "Semaphore"
      summary: "Create a semaphore."
      description: "Creates a Semaphore with the given number of permits"
      operationId: "creatSemaphore"
      produces:
      - "application/json"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - in: "body"
        name: "capacity"
        description: "number of permits"
        required: true
        schema:
          $ref: "#/definitions/capacity"
      responses:
        200:
          description: "semaphore status"
          schema:
            $ref: "#/definitions/Semaphore"
      x-swagger-router-controller: "Semaphore"
  /semaphore/{semaphoreKey}/{semaphoreHandle}:
    get:
      tags:
      - "Semaphore"
      summary: "Query semaphore status"
      description: "Returns the current status in this semaphore."
      operationId: "availablePermits"
      produces:
      - "application/json"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "semaphoreHandle"
        in: "path"
        description: "check semaphore owner"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "semaphore status"
          schema:
            $ref: "#/definitions/Semaphore"
      x-swagger-router-controller: "Semaphore"
    post:
      tags:
      - "Semaphore"
      summary: "Acquires a permit from this semaphore"
      description: "Acquires a permit, reducing the number of available permits by\
        \ one"
      operationId: "acquire"
      produces:
      - "application/json"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "semaphoreHandle"
        in: "path"
        description: "check semaphore owner"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "semaphore status"
          schema:
            $ref: "#/definitions/Semaphore"
      x-swagger-router-controller: "Semaphore"
    put:
      tags:
      - "Semaphore"
      summary: "releases a permit, returning it to the semaphore."
      description: "releases a permit, increasing the number of available permits\
        \ by one"
      operationId: "releases"
      produces:
      - "application/json"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "semaphoreHandle"
        in: "path"
        description: "check semaphore owner"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "semaphore status"
          schema:
            $ref: "#/definitions/Semaphore"
      x-swagger-router-controller: "Semaphore"
    delete:
      tags:
      - "Semaphore"
      summary: "delete semaphore"
      description: "delete semaphore"
      operationId: "deleteSemaphore"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "semaphoreHandle"
        in: "path"
        description: "check semaphore owner"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Semaphore"
    patch:
      tags:
      - "Semaphore"
      summary: "update ttl"
      description: "update ttl"
      operationId: "extendttl"
      parameters:
      - name: "semaphoreKey"
        in: "path"
        description: "semaphore identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "semaphoreHandle"
        in: "path"
        description: "check semaphore owner"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - in: "body"
        name: "ttl"
        description: "update ttl"
        required: true
        schema:
          $ref: "#/definitions/ttl_1"
      responses:
        200:
          description: "semaphore status"
          schema:
            $ref: "#/definitions/Semaphore"
      x-swagger-router-controller: "Semaphore"
  /mutex/{mutexKey}:
    get:
      tags:
      - "Mutex"
      summary: "Query mutex status"
      operationId: "queryMutex"
      produces:
      - "application/json"
      parameters:
      - name: "mutexKey"
        in: "path"
        description: "Mutex identifier"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Mutex"
        400:
          description: "Invalid ID supplied"
      x-swagger-router-controller: "Mutex"
    put:
      tags:
      - "Mutex"
      summary: "Acquires the lock"
      operationId: "lockMutex"
      parameters:
      - name: "mutexKey"
        in: "path"
        description: "Mutex ID"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - in: "body"
        name: "ttl"
        required: false
        schema:
          $ref: "#/definitions/ttl"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Mutex"
        409:
          description: "lock in use"
          schema:
            $ref: "#/definitions/Mutex"
      x-swagger-router-controller: "Mutex"
  /mutex/{mutexKey}/{mutexHandle}:
    delete:
      tags:
      - "Mutex"
      summary: "Releases the lock"
      operationId: "unlockMutex"
      produces:
      - "application/json"
      parameters:
      - name: "mutexKey"
        in: "path"
        description: "Mutex ID"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "mutexHandle"
        in: "path"
        description: "Mutex id to delete"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Mutex"
      x-swagger-router-controller: "Mutex"
    patch:
      tags:
      - "Mutex"
      summary: "Update ttl"
      operationId: "extendMutex"
      parameters:
      - name: "mutexKey"
        in: "path"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - name: "mutexHandle"
        in: "path"
        description: "Mutex id to delete"
        required: true
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      - in: "body"
        name: "mutex"
        required: false
        schema:
          $ref: "#/definitions/ttl"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Mutex"
        400:
          description: "handle mismach"
        410:
          description: "Wrong / expired handle"
      x-swagger-router-controller: "Mutex"
definitions:
  Semaphore:
    type: "object"
    required:
    - "capacity"
    - "count"
    - "expiry"
    - "handle"
    - "id"
    properties:
      id:
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      handle:
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      capacity:
        type: "integer"
        format: "int64"
      count:
        type: "integer"
        format: "int64"
      expiry:
        type: "integer"
        format: "int64"
    example:
      handle: "handle"
      id: "id"
      expiry: 1532486984
      count: 0
      capacity: 10
  Mutex:
    type: "object"
    required:
    - "expiry"
    - "id"
    - "locked"
    properties:
      id:
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      handle:
        type: "string"
        pattern: "^[a-zA-Z0-9\\-_]+$"
      locked:
        type: "boolean"
        default: false
      expiry:
        type: "integer"
        format: "int64"
    example:
      handle: "handle"
      id: "id"
      expiry: 1532486984
      locked: false
  ttl:
    type: "object"
    properties:
      ttl:
        type: "integer"
        minimum: 1
        maximum: 3600
  capacity:
    type: "object"
    properties:
      capacity:
        type: "integer"
        minimum: 1
        maximum: 10
  ttl_1:
    type: "object"
    properties:
      ttl:
        type: "integer"
        minimum: 1
        maximum: 3600